moved ModuleGroup and Snapshot to commons project
