+++ module-array/src/main/java/ch/hsr/adv/ui/array/logic/ArrayDefaultLayouter.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.array.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.ui.array.logic.domain.ArrayElement;
-import ch.hsr.adv.ui.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.presentation.widgets.AutoScalePane;
 import ch.hsr.adv.ui.core.presentation.widgets.LabeledNode;
 import com.google.inject.Singleton;
