+++ module-graph/src/test/java/ch/hsr/adv/ui/graph/logic/GraphParserTest.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.graph.logic;
 
 import ch.hsr.adv.commons.core.logic.domain.ADVElement;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.access.FileDatastoreAccess;
-import ch.hsr.adv.ui.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.logic.util.ADVParseException;
 import ch.hsr.adv.ui.graph.logic.domain.GraphElement;
 import com.google.gson.Gson;
