+++ module-graph/src/test/java/ch/hsr/adv/ui/graph/logic/GraphStringifyerTest.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.ui.graph.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.access.FileDatastoreAccess;
-import ch.hsr.adv.ui.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.logic.util.ADVParseException;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
