+++ ui-core/src/main/java/ch/hsr/adv/ui/core/logic/ADVFlowControl.java
@@ -1,9 +1,9 @@
 package ch.hsr.adv.ui.core.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.Snapshot;
 import ch.hsr.adv.ui.core.access.DatastoreAccess;
 import ch.hsr.adv.ui.core.logic.domain.LayoutedSnapshot;
 import ch.hsr.adv.ui.core.logic.domain.Session;
-import ch.hsr.adv.ui.core.logic.domain.Snapshot;
 import ch.hsr.adv.ui.core.logic.events.ADVEvent;
 import ch.hsr.adv.ui.core.logic.events.EventManager;
 import ch.hsr.adv.ui.core.logic.stores.LayoutedSnapshotStore;
