+++ ui-core/src/main/java/ch/hsr/adv/ui/core/logic/stores/SessionStore.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.ui.core.logic.stores;
 
+import ch.hsr.adv.commons.core.logic.domain.Snapshot;
 import ch.hsr.adv.ui.core.logic.domain.Session;
-import ch.hsr.adv.ui.core.logic.domain.Snapshot;
 import ch.hsr.adv.ui.core.logic.events.ADVEvent;
 import ch.hsr.adv.ui.core.logic.events.EventManager;
 import org.slf4j.Logger;
