+++ ui-core/src/test/java/ch/hsr/adv/ui/core/logic/mocks/TestParser.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.core.logic.mocks;
 
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.logic.GsonProvider;
 import ch.hsr.adv.ui.core.logic.Parser;
-import ch.hsr.adv.ui.core.logic.domain.ModuleGroup;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
