+++ ui-core/src/test/java/ch/hsr/adv/ui/core/logic/mocks/TestStringifyer.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.ui.core.logic.mocks;
 
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.logic.Stringifyer;
-import ch.hsr.adv.ui.core.logic.domain.ModuleGroup;
 import com.google.gson.JsonElement;
 import com.google.inject.Singleton;
